// angular/src/app/shared/postcode-epc-service/model/response/epc-response.ts
export interface EpcResponse {
    'lmk-key': string;
    'address1': string;
    'address2': string;
    'address3': string;
    'postcode': string;
    'building-reference-number': string;
    'current-energy-rating': string;
    'potential-energy-rating': string;
    'current-energy-efficiency': string;
    'potential-energy-efficiency': string;
    'property-type': string;
    'built-form': string;
    'inspection-date': string;
    'local-authority': string;
    'constituency': string;
    'county': string;
    'lodgement-date': string;
    'transaction-type': string;
    'environment-impact-current': string;
    'environment-impact-potential': string;
    'energy-consumption-current': string;
    'energy-consumption-potential': string;
    'co2-emissions-current': string;
    'co2-emiss-curr-per-floor-area': string;
    'co2-emissions-potential': string;
    'lighting-cost-current': string;
    'lighting-cost-potential': string;
    'heating-cost-current': string;
    'heating-cost-potential': string;
    'hot-water-cost-current': string;
    'hot-water-cost-potential': string;
    'total-floor-area': string;
    'energy-tariff': string;
    'mains-gas-flag': string;
    'floor-level': string;
    'flat-top-storey': string;
    'flat-storey-count': string;
    'main-heating-controls': string;
    'multi-glaze-proportion': string;
    'glazed-type': string;
    'glazed-area': string;
    'extension-count': string;
    'number-habitable-rooms': string;
    'number-heated-rooms': string;
    'low-energy-lighting': string;
    'number-open-fireplaces': string;
    'hotwater-description': string;
    'hot-water-energy-eff': string;
    'hot-water-env-eff': string;
    'floor-description': string;
    'floor-energy-eff': string;
    'floor-env-eff': string;
    'windows-description': string;
    'windows-energy-eff': string;
    'windows-env-eff': string;
    'walls-description': string;
    'walls-energy-eff': string;
    'walls-env-eff': string;
    'secondheat-description': string;
    'sheating-energy-eff': string;
    'sheating-env-eff': string;
    'roof-description': string;
    'roof-energy-eff': string;
    'roof-env-eff': string;
    'mainheat-description': string;
    'mainheat-energy-eff': string;
    'mainheat-env-eff': string;
    'mainheatcont-description': string;
    'mainheatc-energy-eff': string;
    'mainheatc-env-eff': string;
    'lighting-description': string;
    'lighting-energy-eff': string;
    'lighting-env-eff': string;
    'main-fuel': string;
    'wind-turbine-count': string;
    'heat-loss-corridoor': string;
    'unheated-corridor-length': string;
    'floor-height': string;
    'photo-supply': string;
    'solar-water-heating-flag': string;
    'mechanical-ventilation': string;
    'address': string;
    'local-authority-label': string;
    'constituency-label': string;
    'certificate-hash': string;
}